Refactor admins.py
